refactor: solution 관련 유틸을 라이브러리로 변경
- getSoltuions 함수 분리 및 디렉터리 이동
- parse 관련 부분 분리
