+++ packages/cli/src/lib/generate/readme.ts
@@ -1,11 +1,11 @@
 import path from 'node:path';
 
 import { createREADMETemplate } from '@/lib/template';
-import { getSolutions } from '@/utils';
 import { ensureWriteFile } from '@/utils/fs';
+import { getSolutionGroupByCodingSite } from '@/api/solution';
 
 export async function generateREADME(solutionDir: string, outDir: string) {
-  const solutions = await getSolutions(solutionDir, outDir);
+  const solutionGroups = await getSolutionGroupByCodingSite(solutionDir, outDir);
 
-  await ensureWriteFile(path.join(outDir, 'README.md'), createREADMETemplate(solutions));
+  await ensureWriteFile(path.join(outDir, 'README.md'), createREADMETemplate(solutionGroups));
 }
