+++ packages/cli/src/lib/parse/index.ts
@@ -1,2 +1,3 @@
 export { parseLeetCode } from './leetcode';
 export { parseProgrammers } from './programmers';
+export { parseSolution } from './solution';
