--- packages/cli/src/utils/index.ts
@@ -1,1 +1,0 @@
-export * from './getSolutions';
